+++ src/borg/invert_command.rs
@@ -1,3 +1,4 @@
+///! Extract information from previous borg run command lines
 use crate::config;
 use std::collections::BTreeSet;
 use std::path::PathBuf;